+++ src/Parser.php
@@ -54,7 +54,7 @@ namespace SimplePie;
  * @package SimplePie
  * @subpackage Parsing
  */
-class Parser
+class Parser implements RegistryAware
 {
     public $error_code;
     public $error_string;
@@ -73,7 +73,7 @@ class Parser
     public $encoding;
     protected $registry;
 
-    public function set_registry(\SimplePie\Registry $registry)
+    public function set_registry(\SimplePie\Registry $registry)/* : void */
     {
         $this->registry = $registry;
     }